Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Is there any valid reason to update the build fingerprint? Is this new one certified?

Yes, the valid reason to update the build fingerprint is to get past the safetynet check (specifically the cts profile) being certified. This cts profile check is not certified with the current build fingerprint. I have checked the new build fingerprint with the hideprops magisk mudule. The cts profile check (and all other safetynet checks) are certified with the new build fingerprint.

Patch-set: 2
CC: Gerrit User 22509 <22509@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>